requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
googlesearch-python>=1.2.3
pandas>=2.0.0
openpyxl>=3.1.0
undetected-chromedriver>=3.5.4
pymysql>=1.1.0